+++ AI5R-STUDIO/dashboard/src/design-system/docking/index.js
@@ -1,0 +1,4 @@
+export { default as DockManager } from "./DockManager";
+export { default as DockRegistry, DOCK_AREAS } from "./DockRegistry";
+export { default as DockLayout } from "./DockLayout";
+export { default as DockPanel } from "./DockPanel";